requirements.txt
selenium==4.18.1
playwright==1.44.0
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.1
webdriver-manager==4.0.1
google-auth==2.29.0
google-api-python-client==2.126.0
